Completed gap jumping mechanism
Gap jumping: i noticed that there are some gaps in the riot skin id's.
for example in the simple way of download, we go until the 403, wich
means the end of the skins. at MasterYi, we get the error after id 6,
but there are 2 skins with id 11, and 12.

So, created an enum map at Champions, with gap size list.
(Also, found 5 champs, where are more than one gap exist)

Now, if we run into a 403 response, we try to jump with the next gap
number on the list.